+++ frontend/backup/main copy.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+import 'package:techtags/screens/navigation_menu.dart';
+
+void main() {
+  runApp(MaterialApp(
+      title: 'TecTags: Object Counting App',
+      debugShowCheckedModeBanner: false, // Removes the "Demo" banner
+      home: NavigationMenu()));
+}
