+++ frontend/lib/main.dart
@@ -1,9 +1,59 @@
 import 'package:flutter/material.dart';
-import 'package:image_picker_demo/screens/navigation_menu.dart';
+import 'package:techtags/screens/splash_screen.dart';
+import 'package:techtags/theme/theme.dart';
 
 void main() {
-  runApp(MaterialApp(
-      title: 'TecTags: Object Counting App',
-      debugShowCheckedModeBanner: false, // Removes the "Demo" banner
-      home: NavigationMenu()));
+  runApp(const MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'TecTags App',
+      theme: lightMode,
+      home: SplashScreen(),
+    );
+  }
+}
+
+class FirstRoute extends StatelessWidget {
+  const FirstRoute({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('First Route')),
+      body: Center(
+        child: ElevatedButton(
+          child: const Text('Open route'),
+          onPressed: () {
+            // Navigate to second route when tapped.
+          },
+        ),
+      ),
+    );
+  }
+}
+
+class SecondRoute extends StatelessWidget {
+  const SecondRoute({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Second Route')),
+      body: Center(
+        child: ElevatedButton(
+          onPressed: () {
+            // Navigate back to first route when tapped.
+          },
+          child: const Text('Go back!'),
+        ),
+      ),
+    );
+  }
 }
