+++ frontend/lib/screens/opencv/opencv.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:image_picker_demo/logic/opencv/opencv_photo_viewer.dart';
+import 'package:techtags/logic/opencv/opencv_photo_viewer.dart';
 import 'dart:io';
 import 'package:screenshot/screenshot.dart';
 import 'package:image_picker/image_picker.dart';
