+++ frontend/lib/widgets/welcome_button.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:tectags_flutter/screens/signup_screen.dart';
+
 
 class WelcomeButton extends StatelessWidget {
   const WelcomeButton(
-    {super.key, this.buttonText, this.onTap, this.color, this.textColor});
+      {super.key, this.buttonText, this.onTap, this.color, this.textColor});
   final String? buttonText;
   final Widget? onTap;
   final Color? color;
@@ -14,28 +14,28 @@ class WelcomeButton extends StatelessWidget {
     return GestureDetector(
       onTap: () {
         Navigator.push(
-          context, 
+          context,
           MaterialPageRoute(
             builder: (e) => onTap!,
-            ),
-          );
+          ),
+        );
       },
       child: Container(
         padding: const EdgeInsets.all(30.0),
         decoration: BoxDecoration(
-        color: color!,
-        borderRadius: const BorderRadius.only(
-          topLeft: Radius.circular(50),
+          color: color!,
+          borderRadius: const BorderRadius.only(
+            topLeft: Radius.circular(50),
+          ),
         ),
-      ),
-      child: Text(
-        buttonText!,
-        textAlign: TextAlign.center,
-        style: TextStyle(
-          fontSize: 20.0,
-          fontWeight: FontWeight.bold,
-          color: textColor!,
-         ),
+        child: Text(
+          buttonText!,
+          textAlign: TextAlign.center,
+          style: TextStyle(
+            fontSize: 20.0,
+            fontWeight: FontWeight.bold,
+            color: textColor!,
+          ),
         ),
       ),
     );
